+++ src/common/ErrorCounter.ts
